Fix installation via Composer
As you don't have any tagged versions yet, the current command does not work when the global minimum stability of the user is `stable` (which is the default.) So, the only way to get this working for everyone is to force the installation of `dev-master` and install the oauth2 dep first. As soon as you release a stable version, you will be able to simplify the install command.