Trigger patch in platformsh/client package to actually apply
It seems cweagans/composer-patches has a bug which means patches are not
re-applied to dependencies of dependencies that are already cached by
Composer. Adding this empty patches object for the relevant package appears
to work around the problem.

The patch in question addresses the problem of an access token refresh (an
OAuth2 API call) being triggered unnecessarily, e.g. at the end of a command
run. Before this commit, the patch could have been in previous releases or
not, depending on the cache at the time of packages being built.