Use primary route URL from deployments API in Drush aliases
Now Drush aliases have the same URL that is used in the 'platform drush'
command. The deployment routes are filtered to only the 'upstream' routes that
match the app, then one is selected, preferring the primary route, and then
preferring HTTPS routes, and then preferring short URLs. If no route is found
the fallback is the environment's 'public-url' HAL link.

There could be a performance hit for new or updated environments as the
deployment resource will be fetched as well as the environment. But the API is
quite quick when a connection is already open, and deployments are cached
against the environment HEAD commit.

(the ugly part is that this introduces is another dependency on the Api service)