Allow many commands to work 'locally' on the container
It's now possible to use the following commands on the Platform.sh container
without needing to have an API token or log in to the API:

* db:dump
* db:sql (sql)
* service:redis-cli (redis)
* service:mongo:dump (mongodump)
* service:mongo:export (mongoexport)
* service:mongo:restore (mongorestore)
* service:mongo:shell (mongo)

N.B.: Commands that need to read schema information in some cases (e.g. db:dump)
will not work locally without either providing the --schema option, or logging
in to use the API and then selecting the schema when prompted.