Fix performance regression with many environments and Drush aliases
Drush aliases now depend on the deployments API for the site URL. However,
fetching this information for every environment, whenever the environments list
is refreshed, is very slow for projects with many environments.

This commit ensures that while "drush-aliases --recreate" will fetch the
deployment for every environment, other commands will not.