Refine description of --yes and --no
Symfony Console traditionally has a --no-interaction (-n) option. We don't use
it here since it can be a little dangerously confused with the --no (-n) of
Drush. The --yes and --no options are the replacement: both disable interaction,
and both have a similar result except in the specific case of yes/no
confirmation questions.