Remove drush-aliases SSH calls to find the absolute app root (#869)
This was done for dedicated (Enterprise) environments (whose absolute app root
was not /app). Drush now supports relative paths (since 9.6.2) so this is
unnecessary.

Users whose dedicated (Enterprise) sites are on old versions of Drush 9 (i.e.
between 9.0.0 and 9.6.2) are advised to upgrade Drush to at least 9.6.2. 9.7.1
is the latest version at the time of writing.

See the Drush fix in https://github.com/drush-ops/drush/pull/4019