Use Docker credential helpers to store secrets
There are 3 Docker credential helpers available from:
https://github.com/docker/docker-credential-helpers

These are CLI application binaries (written in Go) which allow storing secrets
in a system appropriate way: Keychain on OS X, the libsecret keyring on Linux
(the GNOME keyring), and the Credential Manager on Windows.

This falls back to the previous file-based storage for systems incompatible with
the credential helpers.