Avoid sending prompt=select_account to authorization when not logged in
The prompt=select_account parameter forces a new account selection to be made on
the authentication provider, and therefore password and 2FA input. This behavior
should only be necessary in these conditions:

1. when the user wants to switch account
2. when the user wants to log in "again" to refresh 2FA status

During the 'login_required' event, as in this commit, the user has made no such
indication. Passing the parameter is unhelpful and prevents the user from
reusing their browser session.