Fix edge case bug
It is possible in an edge situation that the --cleanup flag proposes to ALTER the information_schema tables. Since those cannot be OPTIMIZED, or ALTERED this isn't supposed to happen.

I propose to add a TABLE_TYPE="base table" (as per docs : https://dev.mysql.com/doc/refman/5.7/en/information-schema-tables-table.html ) so that it only lists actual tables, and never SYSTEM VIEW tables.