Stop the user in case they add resources that exceed their limit
Since the current API does not provide a way to stop an update, this might lead in a state where the environment is not working and even doing forks will initialize new environment in this broken state.

This PR tries to calculate if the requested update will exceed the current user limit and if it does, it stops the operation. In case the user wants to proceed, they can use the `--force` argument.