Only use the SSH certificate if available
Adds an `ssh.cert_only` config option which if enabled means only the
certificate will be included in config, if a certificate exists, and not the
user's "default" public key files (~/.ssh/id_rsa etc.). This is disabled by
default. It can be enabled in the config file or in the environment (e.g.
`export {PREFIX}_SSH_CERT_ONLY=1`).

Rename api.ssh_cert_only to ssh.cert_only