Add --resources-init option to five more commands
If the api.sizing configuration flag is enabled, this adds the option to the
following affected commands, in addition to the existing option on the
`environment:push` (push) command. The option can only be used with projects
that support the flexible resources API. It sets the strategy to be used for
picking the size of new resources on the target environment: `parent`, `child`,
`default`, `minimum`, or `manual`.

Affected commands:

* `backup:restore`
* `environment:activate`
* `environment:branch` (`branch`)
* `environment:merge` (`merge`)
* `environment:synchronize` (`sync`)