Add a confirm step before creating a backup (#1414)
This confirm step is added because:

* The `backup` command has been run inadvertently by users who meant to run `backups` (to list backups).
* It can be surprisingly destructive: a backup will delete an older backup if the quota has been reached.

The above reasons are considered to outweigh a potential, but unlikely, backwards compatibility break. Scripts may block indefinitely after this change if they erroneously run `platform backup` in interactive mode. The solution is to disable interactive mode by adding the flag `--yes` or `-y` or `--no-interaction`, or by setting the environment variable `PLATFORMSH_CLI_NO_INTERACTION=1`.