Support listing teams for a single project
The `team:list` (`teams`) command will now filter the teams list to those with
access to the selected project, if any. The project is selected in the normal
way, e.g. by `--project` (`-p`), or the current Git repository.

Use the `--all` option to list all the teams in the organization.

Add the `granted_at` column (`--columns +granted_at`) to see when the team was
added to the project.